+++ tests/Authorization/AuthorizationIntegrationTest.php
@@ -19,15 +19,15 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
         $form = $this->formalizeService->addForm($form);
         // manage action:
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
-            [ResourceActionGrantService::MANAGE_ACTION]));
+            ResourceActionGrantService::MANAGE_ACTION));
 
         $this->formalizeService->removeForm($form);
         // manage action:
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
-            [ResourceActionGrantService::MANAGE_ACTION]));
+            ResourceActionGrantService::MANAGE_ACTION));
     }
 
     public function testSubmissionLevelAuthorizationAddAndRemoveSubmissionResource()
@@ -41,15 +41,15 @@ class AuthorizationIntegrationTest extends AbstractTestCase
         $submission = $this->formalizeService->addSubmission($submission);
 
         // manage action:
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
-            [ResourceActionGrantService::MANAGE_ACTION]));
+            ResourceActionGrantService::MANAGE_ACTION));
 
         $this->formalizeService->removeSubmission($submission);
         // manage action:
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
-            [ResourceActionGrantService::MANAGE_ACTION]));
+            ResourceActionGrantService::MANAGE_ACTION));
     }
 
     public function testNoSubmissionLevelAuthorizationAddAndRemoveSubmissionResource()
@@ -63,9 +63,9 @@ class AuthorizationIntegrationTest extends AbstractTestCase
         $submission = $this->formalizeService->addSubmission($submission);
 
         // manage action:
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
-            [ResourceActionGrantService::MANAGE_ACTION]));
+            ResourceActionGrantService::MANAGE_ACTION));
     }
 
     public function testSubmissionLevelAuthorizationRemoveAllSubmissionResourcesOnRequest(): void
@@ -85,26 +85,34 @@ class AuthorizationIntegrationTest extends AbstractTestCase
         $submission3->setForm($form);
         $submission3 = $this->formalizeService->addSubmission($submission3);
 
-        $this->assertCount(3, $this->testEntityManager->getEntityManager()->getRepository(Submission::class)
+        $this->assertCount(3,
+            $this->testEntityManager->getEntityManager()->getRepository(Submission::class)
             ->findBy(['form' => $form->getIdentifier()]));
 
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
 
         $this->formalizeService->removeAllSubmittedFormSubmissions($form->getIdentifier());
 
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
 
-        $this->assertCount(0, $this->testEntityManager->getEntityManager()->getRepository(Submission::class)
+        $this->assertCount(0,
+            $this->testEntityManager->getEntityManager()->getRepository(Submission::class)
             ->findBy(['form' => $form->getIdentifier()]));
     }
 
@@ -129,21 +137,27 @@ class AuthorizationIntegrationTest extends AbstractTestCase
         $submission3->setForm($form);
         $submission3 = $this->formalizeService->addSubmission($submission3);
 
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
 
         $this->formalizeService->removeForm($form);
 
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedAnyOfItemActions(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(), [ResourceActionGrantService::MANAGE_ACTION]));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
+            ResourceActionGrantService::MANAGE_ACTION));
     }
 
     private function addFormWithSubmissionLevelAuthorization(bool $grantBasedSubmissionAuthorization = true): Form