+++ tests/TestSubmissionEventSubscriber.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Dbp\Relay\FormalizeBundle\Tests;
 
 use Dbp\Relay\FormalizeBundle\Event\CreateSubmissionPostEvent;
+use Dbp\Relay\FormalizeBundle\Event\SubmissionGrantAddedEvent;
+use Dbp\Relay\FormalizeBundle\Event\SubmissionSubmittedPostEvent;
 use Dbp\Relay\FormalizeBundle\Event\SubmittedSubmissionUpdatedPostEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -12,18 +14,30 @@ class TestSubmissionEventSubscriber implements EventSubscriberInterface
 {
     private bool $wasOnCreateSubmissionPostEventCalled = false;
     private bool $wasOnUpdateSubmissionPostEventCalled = false;
+    private bool $wasOnSubmissionSubmittedPostEventCalled = false;
+    private ?SubmissionGrantAddedEvent $submissionGrantAddedEvent = null;
 
     public static function getSubscribedEvents(): array
     {
         return [
             CreateSubmissionPostEvent::class => 'onCreateSubmissionPostEvent',
+            SubmissionSubmittedPostEvent::class => 'onSubmissionSubmittedPostEvent',
             SubmittedSubmissionUpdatedPostEvent::class => 'onUpdateSubmissionPostEvent',
+            SubmissionGrantAddedEvent::class => 'onSubmissionGrantAddedEvent',
         ];
     }
 
+    /**
+     * @deprecated
+     */
     public function onCreateSubmissionPostEvent(CreateSubmissionPostEvent $event): void
     {
         $this->wasOnCreateSubmissionPostEventCalled = true;
+    }
+
+    public function onSubmissionSubmittedPostEvent(SubmissionSubmittedPostEvent $event): void
+    {
+        $this->wasOnSubmissionSubmittedPostEventCalled = true;
     }
 
     public function onUpdateSubmissionPostEvent(SubmittedSubmissionUpdatedPostEvent $event): void
@@ -31,9 +45,22 @@ class TestSubmissionEventSubscriber implements EventSubscriberInterface
         $this->wasOnUpdateSubmissionPostEventCalled = true;
     }
 
+    public function onSubmissionGrantAddedEvent(SubmissionGrantAddedEvent $event): void
+    {
+        $this->submissionGrantAddedEvent = $event;
+    }
+
+    /**
+     * @deprecated
+     */
     public function wasCreateSubmissionPostEventCalled(): bool
     {
         return $this->wasOnCreateSubmissionPostEventCalled;
+    }
+
+    public function wasSubmissionSubmittedPostEventCalled(): bool
+    {
+        return $this->wasOnSubmissionSubmittedPostEventCalled;
     }
 
     public function wasUpdateSubmissionPostEventCalled(): bool
@@ -41,9 +68,16 @@ class TestSubmissionEventSubscriber implements EventSubscriberInterface
         return $this->wasOnUpdateSubmissionPostEventCalled;
     }
 
+    public function getSubmissionGrantAddedEvent(): ?SubmissionGrantAddedEvent
+    {
+        return $this->submissionGrantAddedEvent;
+    }
+
     public function reset(): void
     {
         $this->wasOnCreateSubmissionPostEventCalled = false;
+        $this->wasOnSubmissionSubmittedPostEventCalled = false;
         $this->wasOnUpdateSubmissionPostEventCalled = false;
+        $this->submissionGrantAddedEvent = null;
     }
 }